Moved app.jsx to components
