--- src/App.css
@@ -1,27 +1,0 @@
-.content {
-   width: 1280px;
-   margin: 0 auto;;
-}
-
-*,
-html * /* override x.xhtml.ru style */ 
-{
-  scrollbar-width: thin;
-  scrollbar-color: #8585AD #2F2F37;
-}
-
-*::-webkit-scrollbar,
-html *::-webkit-scrollbar {
-  height: 12px;
-  width: 12px;
-}
-*::-webkit-scrollbar-track,
-html *::-webkit-scrollbar-track {
-  background: #2F2F37;
-}
-*::-webkit-scrollbar-thumb,
-html *::-webkit-scrollbar-thumb {
-  background-color: #8585AD;
-  border-radius: 5px;
-  border: 3px solid #2F2F37;
-}
