+++ src/components/app/app.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 
-import AppHeader from "./components/app-header/app-header";
-import BurgerIngredients from "./components/burger-ingredients/burger-ingredients";
-import BurgerConstructor from "./components/burger-constructor/burger-constructor";
+import AppHeader from "../app-header/app-header";
+import BurgerConstructor from "../burger-constructor/burger-constructor";
+import BurgerIngredients from "../burger-ingredients/burger-ingredients";
 
-import { INGREDIENT_TYPES } from "./constants";
+import { INGREDIENT_TYPES } from '../../constants';
+import { getClassName } from '../../utils';
 
-import "./App.css";
+import style from './app.module.css';
 
 class App extends React.Component {
   state = {
@@ -19,7 +20,7 @@ class App extends React.Component {
     return (
       <React.Fragment>
         <AppHeader />
-        <main className="content d-flex justify-content-between">
+        <main className={getClassName(style.content, style.main) }>
           <BurgerIngredients onIngredientClick={this.onIngredientClick} />
           <BurgerConstructor
             selectedIngredients={this.state.selected}
@@ -43,7 +44,7 @@ class App extends React.Component {
       ...this.state,
     };
 
-    if (ingredient.type == INGREDIENT_TYPES.BUN) {
+    if (ingredient.type === INGREDIENT_TYPES.BUN) {
       this.setBuns(ingredient, newState);
     } else {
       newState.selected.push(ingredient);
@@ -66,8 +67,6 @@ class App extends React.Component {
   };
 
   generateIngredientsList(state) {
-
-
     const other = (state.selected || []).filter(
       (item) =>
         item.type === INGREDIENT_TYPES.MAIN ||
