+++ src/components/app/app.module.css
@@ -1,0 +1,9 @@
+.content {
+   width: 1280px;
+   margin: 0 auto;;
+}
+
+.main {
+   display: flex;
+   justify-content: space-between;
+}
