+++ src/styles/style.css
@@ -1,1 +1,23 @@
-@import '_utils.css';
+@import "_utils.css";
+
+*,
+html * {
+  scrollbar-width: thin;
+  scrollbar-color: #8585ad #2f2f37;
+}
+
+*::-webkit-scrollbar,
+html *::-webkit-scrollbar {
+  height: 12px;
+  width: 12px;
+}
+*::-webkit-scrollbar-track,
+html *::-webkit-scrollbar-track {
+  background: #2f2f37;
+}
+*::-webkit-scrollbar-thumb,
+html *::-webkit-scrollbar-thumb {
+  background-color: #8585ad;
+  border-radius: 5px;
+  border: 3px solid #2f2f37;
+}
